+++ web/person-roles/app/components/Table/OfficerChange/index.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { h } from 'vue'
-import type { TableColumn } from '@nuxt/ui'
+import type { TableColumn, BadgeProps } from '@nuxt/ui'
 import type { Row } from '@tanstack/vue-table'
 
 const { t } = useI18n()
@@ -30,6 +30,47 @@ const formTitle = computed(() => {
     return ''
   }
 })
+
+function getOfficerTableBadges(actions: OfficerFormAction[]): BadgeProps[] {
+  if (!actions.length) {
+    return []
+  }
+
+  const t = useNuxtApp().$i18n.t
+
+  const unique = [...new Set(actions)]
+
+  const badgeMap: Record<OfficerTableBadgeKey, BadgeProps> = {
+    name: {
+      label: t('badge.nameChanged')
+    },
+    roles: {
+      label: t('badge.rolesChanged')
+    },
+    address: {
+      label: t('badge.addressChanged')
+    },
+    added: {
+      label: t('badge.added')
+    },
+    removed: {
+      label: t('badge.removed'),
+      color: 'neutral'
+    }
+  }
+
+  if (unique.includes('removed')) {
+    return [badgeMap['removed']]
+  }
+
+  if (unique.includes('added')) {
+    return [badgeMap['added']]
+  }
+
+  const newBadges = unique.map(i => badgeMap[i])
+
+  return newBadges
+}
 
 function isRowRemoved(row: Row<OfficerTableState>) {
   return row.original.state.actions.includes('removed')
