--- web/person-roles/app/types/officer/index.ts
@@ -1,1 +1,0 @@
-export * from './table'
