+++ web/person-roles/app/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './validate'
-export * from './officer'
