--- web/person-roles/app/utils/officer/index.ts
@@ -1,1 +1,0 @@
-export * from './table-badges'
